+++ scripts/analyze.sh
@@ -9,5 +9,6 @@ if [ ! -f "$TYPES_FILE" ]; then
 fi
 
 luau-lsp analyze --base-luaurc=.luaurc --settings=.luau-analyze.json \
+    --ignore '**/node_modules/**' --ignore 'node_modules/**' \
     --definitions=$TYPES_FILE \
     src