+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Fix `Path.setExtension` implementation to avoid issues when clearing the extension with `setExtension("")` ([#5](https://github.com/seaofvoices/luau-path/pull/5))
+
 ## 0.1.0
 
 - Initial version